requirements.txt
Flask==1.1.2
Flask-MySQL==1.5.2